Replace compiled angular-bootstrap file with original HTML and JS source files.
