+++ .eslintignore
@@ -1,4 +1,4 @@
-src/fixtures
+
 test/fixtures/scenarios
 optimize
 test/fixtures/scenarios
