+++ src/ui/public/angular-bootstrap/accordion/accordion.html
@@ -1,0 +1,1 @@
+<div class="panel-group" ng-transclude></div>
