+++ src/ui/public/angular-bootstrap/carousel/slide.html
@@ -1,0 +1,7 @@
+<div ng-class="{
+    'active': leaving || (active && !entering),
+    'prev': (next || active) && direction=='prev',
+    'next': (next || active) && direction=='next',
+    'right': direction=='prev',
+    'left': direction=='next'
+  }" class="item text-center" ng-transclude></div>
