+++ src/ui/public/angular-bootstrap/modal/backdrop.html
@@ -1,0 +1,4 @@
+<div class="modal-backdrop fade {{ backdropClass }}"
+     ng-class="{in: animate}"
+     ng-style="{'z-index': 1040 + (index && 1 || 0) + index*10}"
+></div>
