+++ src/ui/public/angular-bootstrap/popover/popover.html
@@ -1,0 +1,8 @@
+<div class="popover {{placement}}" ng-class="{ in: isOpen(), fade: animation() }">
+  <div class="arrow"></div>
+
+  <div class="popover-inner">
+      <h3 class="popover-title" ng-bind="title" ng-show="title"></h3>
+      <div class="popover-content" ng-bind="content"></div>
+  </div>
+</div>
