+++ src/ui/public/angular-bootstrap/progressbar/progress.html
@@ -1,0 +1,1 @@
+<div class="progress" ng-transclude></div>
