+++ src/ui/public/angular-bootstrap/tabs/tab.html
@@ -1,0 +1,3 @@
+<li ng-class="{active: active, disabled: disabled}">
+  <a href ng-click="select()" tab-heading-transclude>{{heading}}</a>
+</li>
