+++ src/ui/public/angular-bootstrap/tooltip/tooltip-html-unsafe-popup.html
@@ -1,0 +1,4 @@
+<div class="tooltip {{placement}}" ng-class="{ in: isOpen(), fade: animation() }">
+  <div class="tooltip-arrow"></div>
+  <div class="tooltip-inner" bind-html-unsafe="content"></div>
+</div>
