+++ src/ui/public/angular-bootstrap/tooltip/tooltip-popup.html
@@ -1,0 +1,4 @@
+<div class="tooltip {{placement}}" ng-class="{ in: isOpen(), fade: animation() }">
+  <div class="tooltip-arrow"></div>
+  <div class="tooltip-inner" ng-bind="content"></div>
+</div>
