+++ src/ui/public/angular-bootstrap/typeahead/typeahead-match.html
@@ -1,0 +1,1 @@
+<a tabindex="-1" bind-html-unsafe="match.label | typeaheadHighlight:query"></a>
