+++ webpackShims/ui-bootstrap.js
@@ -1,6 +1,6 @@
 define(function (require) {
   require('angular');
-  require('ui/angular-bootstrap/ui-bootstrap-tpls');
+  require('ui/angular-bootstrap/index');
 
   return require('ui/modules')
   .get('kibana', ['ui.bootstrap'])
